requirements.txt
dependency-injection==1.2.0
dependency-injector==4.41.0
psycopg2-binary==2.9.9
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.28
TgCrypto==1.2.5
transitions==0.9.0
typing_extensions==4.10.0
